PlayerSettings synchronized over the network. You can now exit a live GameSession and start another one. Some other, mostly cosmetic changes.
Player now keeps an instance of PlayerSettings. There's one instance created initially for use by the local player and console.
PlayerView now gets its texture from an ITextureBank<PlayerTextures> instance.
Level now a little fatter, a little harder to fall through (hopefully).
Elapsed time now measured in milliseconds (so the timing values will be less arbitrary...)
IControllers now use floats instead of longs--this will bring it more in sync with how Farseer does things.

NetworkSession cleanup:
INetworkSession now implements IDisposable so that when the inner containers are disposed (when the GameSession ends), the NetworkSession also get disposed.

Did a little bit of work so the game can run in release mode (no real performance difference at the moment).

Added some friction just to see how it goes...

Some minor fixes to how ScreenManager handles "Dead" screens.
MainMenuScreen now checks when its GamePlayScreen exits, and cleans up after it.

Game now runs on a variable timestep. (This has made Farseer much more reliable from what I can tell).

Created a Documents folder and moved our meagre documents there. Also started doing a little research into our eventual networking model, and added some notes.

+++ FreneticGame/Engine/GameConsoleController.cs
@@ -14,7 +14,7 @@ namespace Frenetic
 
         #region IController Members
 
-        public void Process(long ticks)
+        public void Process(float ticks)
         {
             _keyboard.Unlock();
